requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-crispy-forms==1.13.0
django-filter==21.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.1.5
psycopg2==2.9.1
PyJWT==1.7.1
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
typing-extensions==3.10.0.2
xlrd==2.0.1
xlwt==1.3.0
